requirements.txt
# --- Core API ---
fastapi
uvicorn
python-multipart

# --- Numerical & Imaging ---
numpy==1.26.4
Pillow
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84

# --- Vector DB ---
faiss-cpu

# --- Models ---
transformers
ultralytics
fer

# --- Utilities ---
pydantic
tqdm
huggingface-hub

# --- Deep Learning ---
tensorflow-cpu==2.15.0

# Core dependencies (CPU build of PyTorch)
torch==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.20.1+cpu
torchaudio==2.5.1+cpu
